fix gcc warning about pointer-to-int casts

+++ runtime/ref.h
@@ -7,6 +7,6 @@ typedef struct {
 } ref_t;
 
 #define REF_NULL ((ref_t) { .object = NULL, .vtable = NULL })
-#define REF_HASH(ref) ((uint32_t)ref.object)
+#define REF_HASH(ref) ((uint32_t)(intptr_t)ref.object)
 
 #endif // FOO_H_
